requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==3.2
django-cloudinary-storage==0.3.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-summernote==0.8.20.0
django-taggit==2.1.0
idna==3.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
